Move operator directory
