+++ operator/adx.go
