+++ operator/adx_test.go
